new code
i had to delete the old code. i couldnt figure out how to revert my commit, and get my code back to how i wanted it to be before. so after much fingerling and by some miracle getting it back to how i want it to be, i copied the files and folders, deleted the repository and created a new one. :(